moved api specific test utility back to api test.
